feat(inOutItem): 新增对完成项的排除功能及项目选择函数 style(common.less): 调整depot-mask的样式 fix(ImportItemModal): 添加对发票模板的处理 refactor(InvoiceRecord): 更改字段类型为BigDecimal并优化查询逻辑 chore(InvoiceRecord): 删除不必要的接口文件 feat(InvoiceRecordServiceImpl): 新增赋值含税金额发票的功能

+++ jshERP-boot/src/main/java/com/jsh/erp/datasource/entities/InvoiceRecord.java
@@ -8,6 +8,7 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -52,7 +53,7 @@ public class InvoiceRecord implements Serializable {
     /**
      *  开票含税金额
      */
-    private Float taxAmount;
+    private BigDecimal taxAmount;
 
     /**
      *  单据日期
